+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Upcoming
 
 - Added support for `Message.restrictedVisibility` field.
+- Added support for `Member.extraData` field.
 
 ## 9.4.0